Added search and replace functionality
Summary:
Created a SearchAndReplaceDialog component that's used by both
the EditorPage and ArticleEditor. It highlights search results
in yellow with the current result in orange. Individual
occurrences can be replaced on at a time or all occurences can
be replaced at once.

Test Plan:
Run the automated tests which have been created to test this
new feature.

Reviewers: aria

Differential Revision: https://phabricator.khanacademy.org/D16435